refactor: extract rack action bodies from App.svelte (#2025)
Move handleRackContextDuplicate/Delete/Export/Focus into a new
rack-actions.ts module. The long-press and edit/rename interaction
wrappers stay in App.svelte since M14 (#2018/#2075) rewrites their
invocation. Drop now-unused DRAWER_WIDTH and prepareExportQrCode
imports from App.svelte. Behaviour-preserving.
